zoswi-ai-interview/backend/app/middleware/correlation_id.py
import uuid

from fastapi import Request
from starlette.middleware.base import BaseHTTPMiddleware

from app.core.correlation import set_request_id


class CorrelationIdMiddleware(BaseHTTPMiddleware):
    async def dispatch(self, request: Request, call_next):
        request_id = request.headers.get("x-request-id", str(uuid.uuid4()))
        set_request_id(request_id)
        request.state.request_id = request_id
        response = await call_next(request)
        response.headers["X-Request-ID"] = request_id
        return response

