zoswi-ai-interview/backend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.44
alembic==1.14.1
pydantic==2.12.5
pydantic-settings==2.11.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.20
redis==5.2.1
celery==5.4.0
openai==1.74.0
anthropic==0.50.0
tenacity==9.0.0
structlog==25.2.0
httpx==0.28.1
python-dateutil==2.9.0.post0
psycopg[binary]==3.2.13
asyncpg==0.31.0
pypdf==5.4.0
python-docx==1.1.2
email-validator==2.2.0
